feat: Add Arial Light Italic font file (ARIALLGTITL.TTF)
This commit adds the Arial Light Italic TrueType font file
(ARIALLGTITL.TTF) to the repository. This font variant provides
a lighter and italicized version of the standard Arial typeface,
expanding the available font options for users. The addition of
this font aims to improve the design flexibility and aesthetic
appeal within applications utilizing the Arial font family.