feat: Add Arial Medium font file (ArialMdm.ttf)
This commit adds the Arial Medium font file (ArialMdm.ttf) to the assets/fonts/arial directory. This provides the application with the Arial Medium font, enabling its use in text rendering and improving the visual appearance of the user interface.  The addition of this font expands the available font choices for designers and developers.